+++ src/senaite/patient/tests/base.py
@@ -1,0 +1,88 @@
+# -*- coding: utf-8 -*-
+
+import transaction
+import unittest2 as unittest
+from plone.app.testing import PLONE_FIXTURE
+from plone.app.testing import TEST_USER_ID
+from plone.app.testing import FunctionalTesting
+from plone.app.testing import PloneSandboxLayer
+from plone.app.testing import applyProfile
+from plone.app.testing import setRoles
+from plone.testing import zope
+
+
+class SimpleTestLayer(PloneSandboxLayer):
+    """Setup Plone with installed AddOn only
+    """
+    defaultBases = (PLONE_FIXTURE,)
+
+    def setUpZope(self, app, configurationContext):
+        super(SimpleTestLayer, self).setUpZope(app, configurationContext)
+
+        import Products.TextIndexNG3
+        import bika.lims
+        import senaite.core
+        import senaite.app.listing
+        import senaite.impress
+        import senaite.app.spotlight
+        import senaite.patient
+
+        # Load ZCML
+        self.loadZCML(package=Products.TextIndexNG3)
+        self.loadZCML(package=bika.lims)
+        self.loadZCML(package=senaite.core)
+        self.loadZCML(package=senaite.app.listing)
+        self.loadZCML(package=senaite.impress)
+        self.loadZCML(package=senaite.app.spotlight)
+        self.loadZCML(package=senaite.patient)
+
+        # Install product and call its initialize() function
+        zope.installProduct(app, "Products.TextIndexNG3")
+        zope.installProduct(app, "bika.lims")
+        zope.installProduct(app, "senaite.core")
+        zope.installProduct(app, "senaite.app.listing")
+        zope.installProduct(app, "senaite.impress")
+        zope.installProduct(app, "senaite.app.spotlight")
+        zope.installProduct(app, "senaite.patient")
+
+    def setUpPloneSite(self, portal):
+        super(SimpleTestLayer, self).setUpPloneSite(portal)
+        applyProfile(portal, "senaite.core:default")
+        applyProfile(portal, "senaite.patient:default")
+        transaction.commit()
+
+
+###
+# Use for simple tests (w/o contents)
+###
+SIMPLE_FIXTURE = SimpleTestLayer()
+SIMPLE_TESTING = FunctionalTesting(
+    bases=(SIMPLE_FIXTURE, ),
+    name="senaite.storage:SimpleTesting"
+)
+
+
+class SimpleTestCase(unittest.TestCase):
+    layer = SIMPLE_TESTING
+
+    def setUp(self):
+        super(SimpleTestCase, self).setUp()
+
+        self.app = self.layer["app"]
+        self.portal = self.layer["portal"]
+        self.request = self.layer["request"]
+        self.request["ACTUAL_URL"] = self.portal.absolute_url()
+        setRoles(self.portal, TEST_USER_ID, ["LabManager", "Manager"])
+
+
+class FunctionalTestCase(unittest.TestCase):
+    layer = SIMPLE_TESTING
+
+    def setUp(self):
+        super(FunctionalTestCase, self).setUp()
+
+        self.app = self.layer["app"]
+        self.portal = self.layer["portal"]
+        self.request = self.layer["request"]
+        self.request["ACTUAL_URL"] = self.portal.absolute_url()
+        setRoles(self.portal, TEST_USER_ID, ["LabManager", "Member"])