+++ docs/changelog.rst
@@ -4,5 +4,6 @@ Changelog
 1.0.0 (unreleased)
 ------------------
 
+- #8 Added patient workflow and managed permissions
 - #7 Added upgrade step handler for senaite.patient
 - Initial Release